backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
docker>=7.0.0
pydantic>=2.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
cryptography>=42.0.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.2.0
